refactor: 헤더 리렌더링 성능 개선 리팩토링
