+++ src/components/Header.tsx
@@ -1,6 +1,9 @@
 import React from "react";
 
 class Header extends React.Component {
+  shouldComponentUpdate(): boolean {
+    return false;
+  }
   render() {
     return (
       <header className="gnb">
